refactor: split utils into submodules
